+++ lib/arb_merge.dart
@@ -1,4 +1,5 @@
 library arb_merge;
 
 export './src/options.dart';
-export 'src/arb_merge.dart';
+export 'src/arb_merge.dart' show ArbMerge, ArbMergeResult;
+export 'src/file_operations.dart' show FileOperations;